backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-multipart
alipay-sdk-python-all
